+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix inability to read remote chunks of memory bigger than 128 MiB
 - Improve version detection to handle packed Delphi executables
 - Improve IAT search algorithm for Themida/Winlicense 3.x
+- Gracefully handle bitness mismatch between interpreter and target PEs
 
 ## [0.1.0] - 2021-11-13